Drop Python 3.7 from CI
Upon attempting to use the current numpy (1.22.4), GitHub Actions
tried re-build numpy wheels for Python 3.7, and failed. Why?

Seems that numpy 1.22.4 has wheel packages for Python 3.10, 3.9, and
3.8 on PyPI, but not Python 3.7.  So CI attempted to build numpy
wheels on its own.  We need to have a whole bunch of math libraries
installed on CI to have numpy wheels built there, but that probably is
not worth the trouble.

Python 3.7 is fairly old anyway, so let's drop that.